'version2.1'

+++ src/apex_analysis_package/Scrape/scraper.py
@@ -1,7 +1,8 @@
-import requests 
+import requests
 from bs4 import BeautifulSoup
 import pandas as pd
 import re
+
 
 class Scrape:
     def __init__(self, url):
@@ -39,7 +40,7 @@ class Scrape:
         else:
             print(f"Failed to fetch {self.url}")
             return None
-        
+
     def get_df(self):
         """
         Extract legend information and create a DataFrame.
@@ -50,8 +51,8 @@ class Scrape:
         # To extract the first and last page numbers
         legends = self.soup.find_all("div", {"class": "legends-banner__content"})
         legend_list = []  # Initialize a list to store dictionaries for each legend
-        pattern = r'(.+?)(\d.?.?)%(.+)%Avg rank\s(.+)\s\|\slevel\s(\d+)'
-        
+        pattern = r"(.+?)(\d.?.?)%(.+)%Avg rank\s(.+)\s\|\slevel\s(\d+)"
+
         for legend in legends:
             matches = re.match(pattern, legend.text.strip())
 
@@ -64,17 +65,18 @@ class Scrape:
                 avg_level = matches.group(5)
 
                 result_dict = {
-                    'Legend': legend_name,
-                    'Pick Rate': pick_rate,
-                    'Change': change_rate,
-                    'Avg Rank': avg_rank,
-                    'Avg Level': avg_level
+                    "Legend": legend_name,
+                    "Pick Rate": pick_rate,
+                    "Change": change_rate,
+                    "Avg Rank": avg_rank,
+                    "Avg Level": avg_level,
                 }
                 legend_list.append(result_dict)  # Append the dictionary to the list
 
         df = pd.DataFrame(legend_list)
-        df_fixed= self.correct_legend(df)
+        df_fixed = self.correct_legend(df)
         return df_fixed  # Return the list of dictionaries containing legend information
+
     def correct_legend(self, df):
         """
         Corrects data types and formats in the DataFrame.
@@ -85,22 +87,28 @@ class Scrape:
         Returns:
             pd.DataFrame: DataFrame with corrected data types and formats.
         """
-        df['Pick Rate'] = df['Pick Rate'].str.rstrip('%').astype(float)
-        df['Avg Level'] = df['Avg Level'].astype(int)
+        df["Pick Rate"] = df["Pick Rate"].str.rstrip("%").astype(float)
+        df["Avg Level"] = df["Avg Level"].astype(int)
         # Convert 'Change' to integer, considering ▲ as positive and ▼ as negative
-        df['Change'] = df['Change'].str.replace('▲', '').str.replace('▼', '-').str.rstrip('%').astype(float)
+        df["Change"] = (
+            df["Change"]
+            .str.replace("▲", "")
+            .str.replace("▼", "-")
+            .str.rstrip("%")
+            .astype(float)
+        )
 
         # Map 'Avg Rank' to a rank order dictionary
         rank_order = {
-            'Rookie 4': 0,
-            'Rookie 3': 1,
-            'Rookie 2': 2,
-            'Rookie 1': 3,
-            'Bronze 4': 4,
-            'Bronze 3': 5,
-            'Bronze 2': 6,
-            'Bronze 1': 7  # Assuming there is a Bronze 1 level
+            "Rookie 4": 0,
+            "Rookie 3": 1,
+            "Rookie 2": 2,
+            "Rookie 1": 3,
+            "Bronze 4": 4,
+            "Bronze 3": 5,
+            "Bronze 2": 6,
+            "Bronze 1": 7,  # Assuming there is a Bronze 1 level
         }
-        df['Avg Rank'] = df['Avg Rank'].map(rank_order)
-        df = df.sort_values(by='Avg Rank')
+        df["Avg Rank"] = df["Avg Rank"].map(rank_order)
+        df = df.sort_values(by="Avg Rank")
         return df
